backend/requirements.txt
# Web framework and server
fastapi==0.109.2
uvicorn==0.27.1

# Database
motor==3.3.2

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File handling
python-multipart==0.0.9
aiofiles==23.2.1

# Testing
pytest==8.0.2
httpx==0.27.0
pytest-asyncio==0.23.5
python-dotenv==1.0.1

# Additional dependencies
pydantic==2.6.1
pydantic[email]==2.6.1
Pillow==10.2.0 